refactors navigation into a shared library
