+++ app/lib/request.js
@@ -10,6 +10,7 @@ const Exception = require('./exception');
 const Constants = require('./constants');
 const Cache = require('./cache');
 const cache = new Cache(null, Constants.CACHE_OPTIONS);
+const buildURL = require('../../src/navigate/build-url');
 
 class Request {
     /**
@@ -60,17 +61,8 @@ class Request {
     }
 
     static buildURL(path, params) {
-        const validParams = _.omit(params, invalid);
-        const paramString = querystring.stringify(validParams);
-        return `${path}${path[path.length - 1] === '?' ? '' : '?'}${paramString}`;
+        return buildURL(path, params);
     }
-}
-
-function invalid(param) {
-    return (_.isNull(param) ||
-            _.isNaN(param) ||
-            _.isUndefined(param) ||
-            ((_.isArray(param) || _.isString(param)) && _.isEmpty(param)));
 }
 
 module.exports = Request;
